rename: qr scan 모듈 위치 이동
